requirements.txt
docker-compose==1.29.2
paho-mqtt==1.5.1
